[Refactor] Extract stage1 silu/gelu_tanh into activations.py
Move the two elementwise activation helpers out of the compile_moe_gemm1
closure into a module-level activations.py so they can be reused and tested
independently. No behavior change; call sites resolve to the imported helpers.
