+++ kernels/moe/moe_gemm_2stage/gemm1.py
@@ -52,6 +52,7 @@ from kernels.moe.moe_common import (
 from kernels.moe.moe_common import (
     i64x2_to_v8f16 as _i64x2_to_v8f16,
 )
+from kernels.moe.moe_gemm_2stage.activations import gelu_tanh, silu
 
 
 @functools.lru_cache(maxsize=1024)
@@ -318,37 +319,6 @@ def compile_moe_gemm1(
             vec8_elems = 8 if elem_bytes == 1 else 4
             vec8_x = T.vec(vec8_elems, x_elem)
             vec16_x = T.vec(vec16_elems, x_elem)
-
-            def silu(x):
-                # device fast path:
-                #   emu = exp(-x)  ~= exp2(log2e * (-x))  -> v_exp_f32
-                #   sig = rcp(1 + emu)                   -> v_rcp_f32
-                #   y = x * sig
-                #
-                # Using llvm.amdgcn intrinsics prevents lowering to the div_scale/div_fixup
-                # sequences that introduce extra compares/cndmasks.
-                t = x * (-1.4426950408889634)  # -log2(e)
-                emu = rocdl.exp2(T.f32, t)
-                den = 1.0 + emu
-                sig = rocdl.rcp(T.f32, den)
-                return x * sig
-
-            def gelu_tanh(x):
-                # GeLU tanh approx: 0.5*x*(1 + tanh(sqrt(2/pi)*(x + 0.044715*x^3))).
-                # Expand tanh via exp(-2|y|) in [0,1] (non-positive exponent avoids fp32
-                # overflow).
-                half = fx.Float32(0.5)
-                one = fx.Float32(1.0)
-                two = fx.Float32(2.0)
-                zero = fx.Float32(0.0)
-                x3 = x * x * x
-                y = fx.Float32(0.7978845608) * (x + fx.Float32(0.044715) * x3)
-                abs_y = fx.Float32(y).maximumf(zero - y)
-                e = fx.exp(fx.Float32(-2.0) * abs_y)
-                den = one + e
-                # 1 + tanh(y): y>=0 -> 2/den ; y<0 -> 2*e/den
-                numerator = (y > zero).select(two, two * e)
-                return half * x * (numerator * (one / den))
 
             acc_init = arith.constant_vector(0, T.i32x4) if is_int8 else arith.constant_vector(0.0, T.f32x4)
             zero_f32_acc = arith.constant_vector(0.0, T.f32x4) if is_int4_bf16_groupwise else None
